+++ kv-storage/navigation-state.tsx
@@ -1,3 +1,4 @@
+import { getCurrentScreen } from '../navigation/navigation';
 import { storeKv } from './kv-storage';
 
 // The screens 'Profile Option Screen' and 'Search Filter Option Screen'
@@ -14,32 +15,6 @@ const safeScreens = [
   "Home/Traits",
   "Prospect Profile Screen/Prospect Profile",
 ];
-
-const getCurrentScreen = (navigationState: any): string | null => {
-  // Validate the basic structure of the navigation state
-  if (
-    !navigationState ||
-    typeof navigationState !== 'object' ||
-    !Array.isArray(navigationState.routes) ||
-    typeof navigationState.index !== 'number'
-  ) {
-    return null;
-  }
-
-  // Access the current route using the index
-  const currentRoute = navigationState.routes[navigationState.index];
-  if (!currentRoute) {
-    return null;
-  }
-
-  // Recurse into nested state if it exists
-  if (currentRoute.state) {
-    return currentRoute.name + "/" + getCurrentScreen(currentRoute.state);
-  }
-
-  // Return the name of the current route if available
-  return currentRoute.name || null;
-}
 
 const navigationState = async (value?: any) => {
   const currentScreen = getCurrentScreen(value);
